+++ src/roiextractors/extractors/suite2p/suite2psegmentationextractor.py
@@ -23,8 +23,6 @@ class Suite2pSegmentationExtractor(SegmentationExtractor):
     """A segmentation extractor for Suite2p."""
 
     extractor_name = "Suite2pSegmentationExtractor"
-    mode = "folder"
-    installation_mesg = ""  # error message when not installed
 
     @classmethod
     def get_available_channels(cls, folder_path: PathType) -> list[str]:
@@ -78,8 +76,6 @@ class Suite2pSegmentationExtractor(SegmentationExtractor):
         folder_path: PathType,
         channel_name: Optional[str] = None,
         plane_name: Optional[str] = None,
-        combined: Optional[bool] = None,  # TODO: to be removed
-        plane_no: Optional[int] = None,  # TODO: to be removed
     ):
         """Create SegmentationExtractor object out of suite 2p data type.
 
@@ -91,24 +87,7 @@ class Suite2pSegmentationExtractor(SegmentationExtractor):
             The name of the channel to load, to determine what channels are available use Suite2pSegmentationExtractor.get_available_channels(folder_path).
         plane_name: str, optional
             The name of the plane to load, to determine what planes are available use Suite2pSegmentationExtractor.get_available_planes(folder_path).
-
         """
-        if combined:
-            warning_string = "Keyword argument 'combined' is deprecated and will be removed on or after Nov, 2023. "
-            warn(
-                message=warning_string,
-                category=DeprecationWarning,
-            )
-        if plane_no:
-            warning_string = (
-                "Keyword argument 'plane_no' is deprecated and will be removed on or after Nov, 2023 in favor of 'plane_name'."
-                "Specify which stream you wish to load with the 'plane_name' keyword argument."
-            )
-            warn(
-                message=warning_string,
-                category=DeprecationWarning,
-            )
-
         channel_names = self.get_available_channels(folder_path=folder_path)
         if channel_name is None:
             if len(channel_names) > 1:
@@ -153,7 +132,7 @@ class Suite2pSegmentationExtractor(SegmentationExtractor):
         self.options = options.item()
         self._sampling_frequency = self.options["fs"]
         self._num_frames = self.options["nframes"]
-        self._image_size = (self.options["Ly"], self.options["Lx"])
+        self._image_shape = (self.options["Ly"], self.options["Lx"])
 
         self.stat = self._load_npy(file_name="stat.npy", require=True)
 
@@ -253,10 +232,10 @@ class Suite2pSegmentationExtractor(SegmentationExtractor):
             return None
 
         correlation_image = self.options["Vcorr"]
-        if (self.options["yrange"][-1], self.options["xrange"][-1]) == self._image_size:
+        if (self.options["yrange"][-1], self.options["xrange"][-1]) == self._image_shape:
             return correlation_image
 
-        img = np.zeros(self._image_size, correlation_image.dtype)
+        img = np.zeros(self._image_shape, correlation_image.dtype)
         img[
             (self.options["Ly"] - self.options["yrange"][-1]) : (self.options["Ly"] - self.options["yrange"][0]),
             self.options["xrange"][0] : self.options["xrange"][-1],
@@ -290,7 +269,7 @@ class Suite2pSegmentationExtractor(SegmentationExtractor):
         return [pixel_mask[i] for i in roi_idx_]
 
     def get_frame_shape(self) -> tuple[int, int]:
-        return self._image_size
+        return self._image_shape
 
     def get_image_size(self) -> tuple[int, int]:
         warnings.warn(